Return the first running FAR pod
Sometimes when we taint and reboot a node which had FAR, then that pod will be restarting on a different node, and in the meantime the old pod won't be available until it terminates. Therefore, we update GetFenceAgentsRemediationPod to  return the first running pod that match FAR labels. On this scenario GetFenceAgentsRemediationPod will return the new pod, from the other node, the healthy one.