Drop log checks in E2E tests
Drop verification of the existence of the "Success" message in the
controller's logs.

This check is a strong dependency from the implementation, which means
the test might fail in the future just because the log changes (even
just in the AWS fence agent).

Moreover the E2E checks already skip this control when the target node
is the one where FAR resides, which means the other checks are
sufficient for the test to pass.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>