Fix tag version in release document created by CI
Currently the release document created by CI uses github.ref_name in
place of the release version, but since we are not running this job on
tag creation anymore, this variable has the name of the branch instead
than the tag.

This change replaced {{github.ref_name}} with v{{inputs.version}}, which
contains the proper tag version.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>