Modify FAR remediation taint effect from NoExecute to NoSchedule
The remediation taint is aimed to prevent scheduling of workloads on the unhealthy node before the node is considered to be healthy and the reboot took action. Using NoExecute effect when FAR failed to fence the node within those 6 minutes could then lead to a risk of volume inconsistency. In any situation where a pod deletion has not succeeded for 6 minutes, kubernetes will force detach volumes being unmounted if the node is unhealthy at that instant